requirements.txt
pandas==2.1.1
numpy==1.26.3
matplotlib==3.8.2
opencv-contrib-python==4.10.0.82
tk==0.1.0
screeninfo==0.8.1
pillow==10.2.0
scipy==1.14.1